refactor: 정책 동기화 서비스 책임 분리
