+++ src/main/java/com/twojz/y_kit/policy/dto/response/PolicyListResponse.java
@@ -1,10 +1,20 @@
 package com.twojz.y_kit.policy.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.twojz.y_kit.policy.domain.entity.PolicyApplicationEntity;
+import com.twojz.y_kit.policy.domain.entity.PolicyCategoryEntity;
+import com.twojz.y_kit.policy.domain.entity.PolicyCategoryMapping;
+import com.twojz.y_kit.policy.domain.entity.PolicyDetailEntity;
+import com.twojz.y_kit.policy.domain.entity.PolicyEntity;
+import com.twojz.y_kit.policy.domain.entity.PolicyKeywordMapping;
+import com.twojz.y_kit.policy.domain.entity.PolicyQualificationEntity;
+import com.twojz.y_kit.policy.domain.entity.PolicyRegion;
 import io.swagger.v3.oas.annotations.media.Schema;
+import java.util.ArrayList;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.stream.Collectors;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -70,5 +80,81 @@ public class PolicyListResponse {
 
     @Schema(description = "등록일시")
     private LocalDateTime createdAt;
-}
 
+    public static PolicyListResponse from(
+            PolicyEntity entity,
+            PolicyDetailEntity detail,
+            PolicyApplicationEntity application,
+            PolicyQualificationEntity qualification,
+            List<PolicyCategoryMapping> categoryMappings,
+            List<PolicyKeywordMapping> keywordMappings,
+            List<PolicyRegion> regions,
+            boolean isBookmarked
+    ) {
+        String largeCategory = null;
+        String mediumCategory = null;
+        if (categoryMappings != null) {
+            for (PolicyCategoryMapping mapping : categoryMappings) {
+                PolicyCategoryEntity category = mapping.getCategory();
+                if (category.getLevel() == 1) {
+                    largeCategory = category.getName();
+                } else if (category.getLevel() == 2) {
+                    mediumCategory = category.getName();
+                }
+            }
+        }
+
+        List<String> keywords = keywordMappings != null ?
+                keywordMappings.stream()
+                        .map(mapping -> mapping.getKeyword().getKeyword())
+                        .collect(Collectors.toList()) :
+                new ArrayList<>();
+
+        List<String> regionNames = regions != null ?
+                regions.stream()
+                        .map(policyRegion -> policyRegion.getRegion().getName())
+                        .collect(Collectors.toList()) :
+                new ArrayList<>();
+
+        return PolicyListResponse.builder()
+                .policyId(entity.getId())
+                .policyNo(entity.getPolicyNo())
+                .policyName(detail != null ? detail.getPlcyNm() : null)
+                .summary(detail != null ? truncate(detail.getPlcyExplnCn()) : null)
+                .largeCategory(largeCategory)
+                .mediumCategory(mediumCategory)
+                .isApplicationAvailable(isApplicationAvailable(application))
+                .applicationStartDate(application != null ? application.getAplyBgngYmd() : null)
+                .applicationEndDate(application != null ? application.getAplyEndYmd() : null)
+                .supervisingInstitution(detail != null ? detail.getSprvsnInstCdNm() : null)
+                .minAge(qualification != null ? qualification.getSprtTrgtMinAge() : null)
+                .maxAge(qualification != null ? qualification.getSprtTrgtMaxAge() : null)
+                .keywords(keywords)
+                .regions(regionNames)
+                .viewCount(entity.getViewCount())
+                .bookmarkCount(entity.getBookmarkCount())
+                .applicationCount(entity.getApplicationCount())
+                .isBookmarked(isBookmarked)
+                .createdAt(entity.getCreatedAt())
+                .build();
+    }
+
+    private static Boolean isApplicationAvailable(PolicyApplicationEntity application) {
+        if (application == null || application.getAplyBgngYmd() == null || application.getAplyEndYmd() == null) {
+            return false;
+        }
+
+        LocalDate today = LocalDate.now();
+        return !today.isBefore(application.getAplyBgngYmd()) && !today.isAfter(application.getAplyEndYmd());
+    }
+
+    private static String truncate(String text) {
+        if (text == null) {
+            return null;
+        }
+        if (text.length() <= 100) {
+            return text;
+        }
+        return text.substring(0, 100) + "...";
+    }
+}
