+++ src/main/java/com/twojz/y_kit/policy/service/PolicySyncService.java
@@ -2,20 +2,14 @@ package com.twojz.y_kit.policy.service;
 
 import com.twojz.y_kit.external.policy.client.YouthPolicyClient;
 import com.twojz.y_kit.external.policy.dto.YouthPolicy;
-import com.twojz.y_kit.policy.domain.dto.PolicyApplicationDto;
-import com.twojz.y_kit.policy.domain.dto.PolicyDetailDto;
-import com.twojz.y_kit.policy.domain.dto.PolicyQualificationDto;
 import com.twojz.y_kit.policy.domain.entity.PolicyApplicationEntity;
-import com.twojz.y_kit.policy.domain.entity.PolicyCategoryEntity;
 import com.twojz.y_kit.policy.domain.entity.PolicyCategoryMapping;
 import com.twojz.y_kit.policy.domain.entity.PolicyDetailEntity;
 import com.twojz.y_kit.policy.domain.entity.PolicyDocumentEntity;
 import com.twojz.y_kit.policy.domain.entity.PolicyEntity;
-import com.twojz.y_kit.policy.domain.entity.PolicyKeywordEntity;
 import com.twojz.y_kit.policy.domain.entity.PolicyKeywordMapping;
 import com.twojz.y_kit.policy.domain.entity.PolicyQualificationEntity;
 import com.twojz.y_kit.policy.domain.entity.PolicyRegion;
-import com.twojz.y_kit.policy.domain.vo.DocumentParsed;
 import com.twojz.y_kit.policy.service.persistence.PolicyApplicationPersistenceService;
 import com.twojz.y_kit.policy.service.persistence.PolicyCategoryPersistenceService;
 import com.twojz.y_kit.policy.service.persistence.PolicyDetailPersistenceService;
@@ -24,32 +18,23 @@ import com.twojz.y_kit.policy.service.persistence.PolicyEntityPersistenceService
 import com.twojz.y_kit.policy.service.persistence.PolicyKeywordPersistenceService;
 import com.twojz.y_kit.policy.service.persistence.PolicyQualificationPersistenceService;
 import com.twojz.y_kit.policy.service.persistence.PolicyRegionPersistenceService;
-import com.twojz.y_kit.region.entity.Region;
-import com.twojz.y_kit.region.repository.RegionRepository;
 import java.time.Duration;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.function.Function;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.StringUtils;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
 public class PolicySyncService {
     private static final int BATCH_SIZE = 100;
-    private final String reg = "\\s*(,|및)\\s*";
 
     private final YouthPolicyClient youthPolicyClient;
     private final PolicyEntityPersistenceService policyEntityPersistenceService;
@@ -60,8 +45,8 @@ public class PolicySyncService {
     private final PolicyCategoryPersistenceService policyCategoryPersistenceService;
     private final PolicyKeywordPersistenceService policyKeywordPersistenceService;
     private final PolicyRegionPersistenceService policyRegionPersistenceService;
-    private final RegionRepository regionRepository;
-    private final PolicyMapper mapper;
+    private final PolicySectionSyncService policySectionSyncService;
+    private final PolicyMappingSyncService policyMappingSyncService;
     private final PolicyAiAnalysisService policyAiAnalysisService;
 
     /**
@@ -93,9 +78,10 @@ public class PolicySyncService {
 
             deactivateMissingPolicies(policies);
 
-            if (!newPolicies.isEmpty()) {
-                processAiAnalysisForNewPolicies(newPolicies);
-            }
+            // AI 분석 호출은 동기화 성능과 외부 API 의존성을 분리하기 위해 임시 비활성화합니다.
+            // if (!newPolicies.isEmpty()) {
+            //     processAiAnalysisForNewPolicies(newPolicies);
+            // }
 
         } catch (Exception e) {
             log.error("정책 동기화 중 오류 발생", e);
@@ -130,8 +116,19 @@ public class PolicySyncService {
         Map<Long, List<PolicyRegion>> regionMappingsMap =
                 policyRegionPersistenceService.findMapByPolicies(policyEntities);
 
-        Map<String, PolicyCategoryEntity> categoryCache = new HashMap<>();
-        Map<String, PolicyKeywordEntity> keywordCache = new HashMap<>();
+        Map<Long, PolicyDetailEntity> detailMap =
+                policyDetailPersistenceService.findMapByPolicies(policyEntities);
+
+        Map<Long, PolicyApplicationEntity> applicationMap =
+                policyApplicationPersistenceService.findMapByPolicies(policyEntities);
+
+        Map<Long, PolicyQualificationEntity> qualificationMap =
+                policyQualificationPersistenceService.findMapByPolicies(policyEntities);
+
+        Map<Long, PolicyDocumentEntity> documentMap =
+                policyDocumentPersistenceService.findMapByPolicies(policyEntities);
+
+        PolicyMappingSyncService.MappingSyncContext mappingContext = policyMappingSyncService.createContext();
 
         for (YouthPolicy apiPolicy : batch) {
             try {
@@ -149,11 +146,17 @@ public class PolicySyncService {
 
                 policy.activate();
 
-                updatePolicy(apiPolicy, policy,
+                ExistingPolicyRelations relations = new ExistingPolicyRelations(
+                        detailMap.get(policy.getId()),
+                        applicationMap.get(policy.getId()),
+                        qualificationMap.get(policy.getId()),
+                        documentMap.get(policy.getId()),
                         categoryMappingsMap.getOrDefault(policy.getId(), Collections.emptyList()),
                         keywordMappingsMap.getOrDefault(policy.getId(), Collections.emptyList()),
-                        regionMappingsMap.getOrDefault(policy.getId(), Collections.emptyList()),
-                        categoryCache, keywordCache);
+                        regionMappingsMap.getOrDefault(policy.getId(), Collections.emptyList())
+                );
+
+                updatePolicy(apiPolicy, policy, relations, mappingContext);
 
             } catch (Exception e) {
                 log.error("정책 저장 실패: {}", apiPolicy.getPlcyNo(), e);
@@ -179,288 +182,28 @@ public class PolicySyncService {
     private boolean updatePolicy(
             YouthPolicy apiPolicy,
             PolicyEntity policy,
-            List<PolicyCategoryMapping> existingCategoryMappings,
-            List<PolicyKeywordMapping> existingKeywordMappings,
-            List<PolicyRegion> existingRegionMappings,
-            Map<String, PolicyCategoryEntity> categoryCache,
-            Map<String, PolicyKeywordEntity> keywordCache) {
-
-        PolicyDetailDto detailReq = mapper.toDetailRequest(apiPolicy);
-        PolicyApplicationDto appReq = mapper.toApplicationRequest(apiPolicy);
-        PolicyQualificationDto qualReq = mapper.toQualificationRequest(apiPolicy);
-
-        boolean detailChanged = updateOrCreateDetail(policy, detailReq);
-        boolean appChanged = updateOrCreateApplication(policy, appReq);
-        boolean qualChanged = updateOrCreateQualification(policy, qualReq);
-        boolean docChanged = updateOrCreateDocument(policy, apiPolicy.getSbmsnDcmntCn());
-
-        boolean categoryChanged = updateCategoryMappings(apiPolicy, policy, existingCategoryMappings, categoryCache);
-        boolean keywordChanged = updateKeywordMappings(apiPolicy, policy, existingKeywordMappings, keywordCache);
-        boolean regionChanged = updateRegionMappings(apiPolicy, policy, existingRegionMappings);
-
-        return detailChanged || appChanged || qualChanged || docChanged ||
-                categoryChanged || keywordChanged || regionChanged;
-    }
-
-    private boolean updateOrCreateDetail(PolicyEntity policy, PolicyDetailDto dto) {
-        PolicyDetailEntity detail = policyDetailPersistenceService.findNullableByPolicy(policy);
-        boolean isNew = detail == null;
-        if (isNew) {
-            detail = PolicyDetailEntity.builder().policy(policy).build();
-        }
-
-        detail.updateFromApi(dto);
-        if (isNew) {
-            policyDetailPersistenceService.save(detail);
-        }
-        return isNew;
-    }
-
-    private boolean updateOrCreateApplication(PolicyEntity policy, PolicyApplicationDto dto) {
-        PolicyApplicationEntity application = policyApplicationPersistenceService.findNullableByPolicy(policy);
-        boolean isNew = application == null;
-        if (isNew) {
-            application = PolicyApplicationEntity.builder().policy(policy).build();
-        }
-
-        application.updateFromApi(dto);
-        if (isNew) {
-            policyApplicationPersistenceService.save(application);
-        }
-        return isNew;
-    }
-
-    private boolean updateOrCreateQualification(PolicyEntity policy, PolicyQualificationDto dto) {
-        PolicyQualificationEntity qualification = policyQualificationPersistenceService.findNullableByPolicy(policy);
-        boolean isNew = qualification == null;
-        if (isNew) {
-            qualification = PolicyQualificationEntity.builder().policy(policy).build();
-        }
-
-        qualification.updateFromApi(dto);
-        if (isNew) {
-            policyQualificationPersistenceService.save(qualification);
-        }
-        return isNew;
-    }
-
-    private boolean updateOrCreateDocument(PolicyEntity policy, String original) {
-        if (isEmptyDocument(original)) return false;
-        PolicyDocumentEntity document = policyDocumentPersistenceService.findNullableByPolicy(policy);
-        boolean isNew = document == null;
-        if (isNew) {
-            document = PolicyDocumentEntity.builder().policy(policy).build();
-        }
-
-        if (isNew || !java.util.Objects.equals(document.getDocumentsOriginal(), original)) {
-            document.updateOriginal(original);
-            DocumentParsed parsed = DocumentPreprocessor.parse(original);
-            document.updateParsed(parsed);
-            policyDocumentPersistenceService.save(document);
-            return true;
-        }
-
-        return false;
-    }
-
-    private boolean isEmptyDocument(String original) {
-        return original == null ||
-                original.trim().isEmpty() ||
-                original.equals("-") ||
-                original.equals("없음");
-    }
-
-    private boolean updateCategoryMappings(
-            YouthPolicy apiPolicy,
-            PolicyEntity policy,
-            List<PolicyCategoryMapping> existingMappings,
-            Map<String, PolicyCategoryEntity> categoryCache) {
-
-        Function<String, Set<String>> parse = txt -> Arrays.stream(txt.split(reg))
-                .map(String::trim)
-                .filter(s -> !s.isEmpty())
-                .collect(Collectors.toCollection(LinkedHashSet::new));
-
-        List<PolicyCategoryMapping> newMappings = new ArrayList<>();
-
-        if (StringUtils.hasText(apiPolicy.getLclsfNm())) {
-            parse.apply(apiPolicy.getLclsfNm()).forEach(name -> {
-                PolicyCategoryEntity main = findOrCreateCategoryWithCache(name, 1, null, categoryCache);
-                newMappings.add(PolicyCategoryMapping.builder()
-                        .policy(policy)
-                        .category(main)
-                        .build());
-            });
-        }
-
-        if (StringUtils.hasText(apiPolicy.getMclsfNm())) {
-            PolicyCategoryEntity mainParent = newMappings.isEmpty() ? null : newMappings.getFirst().getCategory();
-
-            parse.apply(apiPolicy.getMclsfNm()).forEach(name -> {
-                PolicyCategoryEntity sub = findOrCreateCategoryWithCache(name, 2, mainParent, categoryCache);
-                newMappings.add(PolicyCategoryMapping.builder()
-                        .policy(policy)
-                        .category(sub)
-                        .build());
-            });
-        }
-
-        List<Long> existingCategoryIds = existingMappings.stream()
-                .map(m -> m.getCategory().getId())
-                .toList();
-        List<Long> newCategoryIds = newMappings.stream()
-                .map(m -> m.getCategory().getId())
-                .toList();
-
-        if (existingCategoryIds.equals(newCategoryIds)) {
-            return false;
-        }
-
-        policyCategoryPersistenceService.deleteMappings(existingMappings);
-        policyCategoryPersistenceService.saveMappings(newMappings);
-
-        return true;
-    }
-
-    private boolean updateKeywordMappings(
-            YouthPolicy apiPolicy,
-            PolicyEntity policy,
-            List<PolicyKeywordMapping> existingMappings,
-            Map<String, PolicyKeywordEntity> keywordCache) {
-
-        Map<String, PolicyKeywordMapping> existingKeywordMap = existingMappings.stream()
-                .collect(Collectors.toMap(
-                        m -> m.getKeyword().getKeyword(),
-                        m -> m
-                ));
-
-        final Set<String> newKeywords;
-        if (StringUtils.hasText(apiPolicy.getPlcyKywdNm())) {
-            newKeywords = Arrays.stream(apiPolicy.getPlcyKywdNm().split(","))
-                    .map(String::trim)
-                    .filter(kw -> !kw.isEmpty())
-                    .collect(Collectors.toSet());
-        } else {
-            newKeywords = Collections.emptySet();
-        }
-
-        List<PolicyKeywordMapping> toDelete = existingMappings.stream()
-                .filter(m -> !newKeywords.contains(m.getKeyword().getKeyword()))
-                .toList();
-
-        Set<String> toAddKeywords = new HashSet<>(newKeywords);
-        toAddKeywords.removeAll(existingKeywordMap.keySet());
-
-        if (toDelete.isEmpty() && toAddKeywords.isEmpty()) {
-            return false;
-        }
-
-        if (!toDelete.isEmpty()) {
-            toDelete.forEach(m -> m.getKeyword().decreaseUsageCount());
-            policyKeywordPersistenceService.deleteMappings(toDelete);
-        }
-
-        if (!toAddKeywords.isEmpty()) {
-            List<PolicyKeywordMapping> toAdd = toAddKeywords.stream()
-                    .map(kw -> {
-                        PolicyKeywordEntity keyword = findOrCreateKeywordWithCache(kw, keywordCache);
-                        keyword.increaseUsageCount();
-                        return PolicyKeywordMapping.builder()
-                                .policy(policy)
-                                .keyword(keyword)
-                                .build();
-                    })
-                    .toList();
-            policyKeywordPersistenceService.saveMappings(toAdd);
-        }
-
-        return true;
-    }
-
-    private PolicyCategoryEntity findOrCreateCategoryWithCache(
-            String name,
-            int level,
-            PolicyCategoryEntity parent,
-            Map<String, PolicyCategoryEntity> cache
+            ExistingPolicyRelations relations,
+            PolicyMappingSyncService.MappingSyncContext mappingContext
     ) {
-        String cacheKey = name + "_" + level;
-        return cache.computeIfAbsent(cacheKey, key ->
-                policyCategoryPersistenceService.findByNameAndLevel(name, level)
-                        .orElseGet(() -> policyCategoryPersistenceService.save(
-                                PolicyCategoryEntity.builder()
-                                        .name(name)
-                                        .level(level)
-                                        .parent(parent)
-                                        .isActive(true)
-                                        .build()
-                        ))
+        boolean sectionChanged = policySectionSyncService.syncSections(
+                apiPolicy,
+                policy,
+                relations.detail(),
+                relations.application(),
+                relations.qualification(),
+                relations.document()
         );
-    }
 
-    private PolicyKeywordEntity findOrCreateKeywordWithCache(
-            String keywordText,
-            Map<String, PolicyKeywordEntity> cache
-    ) {
-        return cache.computeIfAbsent(keywordText, key ->
-                policyKeywordPersistenceService.findByKeyword(keywordText)
-                        .orElseGet(() -> policyKeywordPersistenceService.save(
-                                PolicyKeywordEntity.builder()
-                                        .keyword(keywordText)
-                                        .usageCount(0)
-                                        .build()
-                        ))
+        boolean mappingChanged = policyMappingSyncService.syncMappings(
+                apiPolicy,
+                policy,
+                relations.categoryMappings(),
+                relations.keywordMappings(),
+                relations.regionMappings(),
+                mappingContext
         );
-    }
 
-    private boolean updateRegionMappings(
-            YouthPolicy apiPolicy,
-            PolicyEntity policy,
-            List<PolicyRegion> existingMappings) {
-
-        Set<String> existingRegionIds = existingMappings.stream()
-                .map(pr -> pr.getRegion().getCode())
-                .collect(Collectors.toSet());
-
-        Set<String> newRegionCodes = new HashSet<>();
-        if (StringUtils.hasText(apiPolicy.getZipCd())) {
-            newRegionCodes = Arrays.stream(apiPolicy.getZipCd().split("[,;]"))
-                    .map(String::trim)
-                    .filter(StringUtils::hasText)
-                    .collect(Collectors.toSet());
-        }
-
-        List<Region> newRegions = regionRepository.findAllByCodeIn(new ArrayList<>(newRegionCodes));
-        Set<String> newRegionIds = newRegions.stream()
-                .map(Region::getCode)
-                .collect(Collectors.toSet());
-
-        List<PolicyRegion> toDelete = existingMappings.stream()
-                .filter(pr -> !newRegionIds.contains(pr.getRegion().getCode()))
-                .toList();
-
-        Set<String> toAddIds = new HashSet<>(newRegionIds);
-        toAddIds.removeAll(existingRegionIds);
-
-        if (toDelete.isEmpty() && toAddIds.isEmpty()) {
-            return false;
-        }
-
-        if (!toDelete.isEmpty()) {
-            policyRegionPersistenceService.deleteAll(toDelete);
-        }
-
-        if (!toAddIds.isEmpty()) {
-            List<PolicyRegion> toAdd = newRegions.stream()
-                    .filter(region -> toAddIds.contains(region.getCode()))
-                    .map(region -> PolicyRegion.builder()
-                            .policy(policy)
-                            .region(region)
-                            .build())
-                    .toList();
-            policyRegionPersistenceService.saveAll(toAdd);
-        }
-
-        return true;
+        return sectionChanged || mappingChanged;
     }
 
     @Transactional
@@ -482,4 +225,14 @@ public class PolicySyncService {
 
         return toDeactivate.size();
     }
+
+    private record ExistingPolicyRelations(
+            PolicyDetailEntity detail,
+            PolicyApplicationEntity application,
+            PolicyQualificationEntity qualification,
+            PolicyDocumentEntity document,
+            List<PolicyCategoryMapping> categoryMappings,
+            List<PolicyKeywordMapping> keywordMappings,
+            List<PolicyRegion> regionMappings
+    ) {}
 }
